add x64 visual studio build target

+++ renderer/win32/win32_software_device.cpp
@@ -91,7 +91,7 @@ void Win32SoftwareDevice::draw_text(const std::string& text, int x, int y)
 
     unique_ptr<wchar_t[]> wcbuf(new wchar_t[text.size() + 1]);
     mbstowcs_s(nullptr, wcbuf.get(), text.size() + 1, text.c_str(), text.size());
-    m_graphics->DrawString(wcbuf.get(), text.size(), &font, p, &hb);
+    m_graphics->DrawString(wcbuf.get(), static_cast<INT>(text.size()), &font, p, &hb);
 }
 
 ///////////////////////////////////////////////////////////////////////////////
